refactor(tui): extract SessionEventHandler from KimiTUI
将 Session Events（39 个方法）和 Background Task Lifecycle（4 个方法）
共 ~960 行代码从 KimiTUI 提取到 SessionEventHandler。包含事件分发、
turn/step 生命周期、thinking/assistant delta 处理、tool call 管理、
compaction 流程、subagent 管理、MCP 状态渲染、background task 事件。

kimi-tui.ts: 5268 → 4316 行 (-952)
