+++ apps/kimi-code/src/tui/kimi-tui.ts
@@ -147,6 +147,7 @@ import { PlatformSelectorComponent } from './components/dialogs/platform-selecto
 import { PermissionSelectorComponent } from './components/dialogs/permission-selector';
 import { QuestionDialogComponent } from './components/dialogs/question-dialog';
 import { SessionPickerComponent, type SessionRow } from './components/dialogs/session-picker';
+import { SessionEventHandler } from './controllers/session-event-handler';
 import { StreamingUIController } from './controllers/streaming-ui';
 import { TasksBrowserController, type TasksBrowserState } from './controllers/tasks-browser';
 import {
@@ -560,12 +561,12 @@ export class KimiTUI {
   private readonly imageStore = new ImageAttachmentStore();
   private readonly fdPath: string | null = detectFdPath();
   private readonly gitLsFilesCache: GitLsFilesCache;
-  private sessionEventUnsubscribe: (() => void) | undefined;
+  sessionEventUnsubscribe: (() => void) | undefined;
   private pendingExit: PendingExit | null = null;
   private cancelInFlight: (() => void) | undefined;
   // Queues editor messages instead of sending or steering them. Used by /init.
   deferUserMessages = false;
-  private aborted = false;
+  aborted = false;
   private terminalFocusTrackingDispose: (() => void) | undefined;
   private terminalThemeTrackingDispose: (() => void) | undefined;
   // Cleanup callbacks for SIGHUP/SIGTERM listeners and stdout/stderr 'error'
@@ -580,6 +581,7 @@ export class KimiTUI {
   // When true, the migration screen is the whole session: run it, then exit.
   private readonly migrateOnly: boolean;
   readonly streamingUI: StreamingUIController;
+  private readonly sessionEventHandler: SessionEventHandler;
   private readonly tasksBrowserController: TasksBrowserController;
 
   public onExit?: (exitCode?: number) => Promise<void>;
@@ -630,6 +632,7 @@ export class KimiTUI {
       }),
     );
     this.streamingUI = new StreamingUIController(this);
+    this.sessionEventHandler = new SessionEventHandler(this);
     this.tasksBrowserController = new TasksBrowserController(this);
     this.setupEditorHandlers();
     this.buildLayout();
@@ -1685,7 +1688,7 @@ export class KimiTUI {
   }
 
   // Sends a queued message after restoring the agent target captured at enqueue time.
-  private sendQueuedMessage(session: Session, item: QueuedMessage): void {
+  sendQueuedMessage(session: Session, item: QueuedMessage): void {
     this.harness.interactiveAgentId = item.agentId ?? MAIN_AGENT_ID;
     this.sendMessageInternal(session, item.text, {
       parts: item.parts,
@@ -2495,970 +2498,15 @@ export class KimiTUI {
   }
 
   // =========================================================================
-  // Session Events
+  // Session Events — delegated to SessionEventHandler
   // =========================================================================
 
-  private startSessionEventSubscription(): void {
-    const session = this.requireSession();
-    const sendQueued = (item: QueuedMessage): void => {
-      this.sendQueuedMessage(session, item);
-    };
-    this.sessionEventUnsubscribe?.();
-    const mcpOAuthOpener = new McpOAuthAuthorizationUrlOpener(openUrl);
-    const { sessionId } = this.state.appState;
-    this.sessionEventUnsubscribe = session.onEvent((event) => {
-      if (this.aborted) return;
-      if (event.sessionId !== sessionId) return;
-      if (event.type === 'tool.progress') {
-        mcpOAuthOpener.handleToolProgress(event);
-      }
-      this.handleEvent(event, sendQueued);
-    });
-    void this.syncMcpServerStatusSnapshot(session);
-  }
+  private startSessionEventSubscription(): void { this.sessionEventHandler.startSubscription(); }
 
-  private async syncMcpServerStatusSnapshot(session: Session): Promise<void> {
-    let servers: readonly McpServerStatusSnapshot[];
-    try {
-      servers = await session.listMcpServers();
-    } catch (error) {
-      if (this.session !== session || this.aborted) return;
-      const message = error instanceof Error ? error.message : String(error);
-      this.showError(`Failed to sync MCP server status: ${message}`);
-      return;
-    }
-    if (this.session !== session || this.state.appState.sessionId !== session.id) return;
-
-    const visible = selectMcpStartupStatusRows(servers);
-    const visibleNames = new Set(visible.map((server) => server.name));
-    for (const server of visible) {
-      if (this.state.renderedMcpServerStatusKeys.has(server.name)) continue;
-      this.renderMcpServerStatus(server);
-    }
-
-    const hidden: McpServerStatusSnapshot[] = [];
-    for (const server of servers) {
-      if (visibleNames.has(server.name)) continue;
-      if (this.state.renderedMcpServerStatusKeys.has(server.name)) continue;
-      this.state.renderedMcpServerStatusKeys.set(server.name, mcpServerStatusKey(server));
-      hidden.push(server);
-    }
-    if (hidden.length > 0) {
-      this.showStatus(
-        formatMcpStartupStatusSummary(hidden, visible.length),
-        this.state.theme.colors.textMuted,
-      );
-    }
-  }
-
-  // Routes an SDK event to the matching TUI state transition.
   private handleEvent(event: Event, sendQueued: (item: QueuedMessage) => void): void {
-    if (this.routeSubagentEvent(event)) {
-      return;
-    }
-
-    if ('turnId' in event && event.turnId !== undefined) {
-      this.state.currentTurnId = String(event.turnId);
-    }
-
-    switch (event.type) {
-      case 'turn.started':
-        this.handleTurnBegin(event);
-        break;
-      case 'turn.ended':
-        this.handleTurnEnd(event, sendQueued);
-        break;
-      case 'turn.step.started':
-        this.handleStepBegin(event);
-        break;
-      case 'turn.step.interrupted':
-        this.handleStepInterrupted(event);
-        break;
-      case 'turn.step.completed':
-        this.handleStepCompleted(event);
-        break;
-      case 'turn.step.retrying':
-        break;
-      case 'tool.progress':
-        this.handleToolProgress(event);
-        break;
-      case 'assistant.delta':
-        this.handleAssistantDelta(event);
-        break;
-      case 'hook.result':
-        this.handleHookResult(event);
-        break;
-      case 'thinking.delta':
-        this.handleThinkingDelta(event);
-        break;
-      case 'tool.call.started':
-        this.handleToolCall(event);
-        break;
-      case 'tool.call.delta':
-        this.handleToolCallDelta(event);
-        break;
-      case 'tool.result':
-        this.handleToolResult(event);
-        break;
-      case 'agent.status.updated':
-        this.handleStatusUpdate(event);
-        break;
-      case 'session.meta.updated':
-        this.handleSessionMetaChanged(event);
-        break;
-      case 'skill.activated':
-        this.handleSkillActivated(event);
-        break;
-      case 'error':
-        this.handleSessionError(event);
-        break;
-      case 'warning':
-        this.handleSessionWarning(event);
-        break;
-      case 'compaction.started':
-        this.handleCompactionBegin(event);
-        break;
-      case 'compaction.completed':
-        this.handleCompactionEnd(event, sendQueued);
-        break;
-      case 'compaction.blocked':
-        break;
-      case 'compaction.cancelled':
-        this.handleCompactionCancel(event, sendQueued);
-        break;
-      case 'subagent.spawned':
-        this.handleSubagentSpawned(event);
-        break;
-      case 'subagent.completed':
-        this.handleSubagentCompleted(event);
-        break;
-      case 'subagent.failed':
-        this.handleSubagentFailed(event);
-        break;
-      case 'background.task.started':
-      case 'background.task.updated':
-      case 'background.task.terminated':
-        this.handleBackgroundTaskEvent(event);
-        break;
-      case 'mcp.server.status':
-        this.renderMcpServerStatus(event.server);
-        break;
-      case 'tool.list.updated':
-        break;
-      default:
-        break;
-    }
+    this.sessionEventHandler.handleEvent(event, sendQueued);
   }
-
-  // Routes child-agent events into their parent tool-call component.
-  private routeSubagentEvent(event: Event): boolean {
-    const subagentId = event.agentId;
-    if (subagentId === MAIN_AGENT_ID) return false;
-
-    const parentToolCallId = this.state.subagentParentToolCallIds.get(subagentId);
-    if (parentToolCallId === undefined || parentToolCallId.length === 0) return true;
-    const sourceName = this.state.subagentNames.get(subagentId);
-    const toolCall = this.state.pendingToolComponents.get(parentToolCallId);
-    if (toolCall === undefined) return true;
-    toolCall.setSubagentMeta(subagentId, sourceName);
-
-    switch (event.type) {
-      case 'hook.result':
-        toolCall.appendSubagentText(formatHookResultPlain(event), 'text');
-        return true;
-      case 'assistant.delta':
-        toolCall.appendSubagentText(event.delta, 'text');
-        return true;
-      case 'thinking.delta':
-        toolCall.appendSubagentText(event.delta, 'thinking');
-        return true;
-      case 'tool.call.started':
-        toolCall.appendSubToolCall({
-          id: `${subagentId}:${event.toolCallId}`,
-          name: event.name,
-          args: argsRecord(event.args),
-        });
-        return true;
-      case 'tool.call.delta':
-        toolCall.appendSubToolCallDelta({
-          id: `${subagentId}:${event.toolCallId}`,
-          name: event.name,
-          argumentsPart: event.argumentsPart ?? null,
-        });
-        return true;
-      case 'tool.result':
-        toolCall.finishSubToolCall({
-          tool_call_id: `${subagentId}:${event.toolCallId}`,
-          output: serializeToolResultOutput(event.output),
-          is_error: event.isError,
-        });
-        return true;
-      case 'agent.status.updated': {
-        const usageObj = event.usage;
-        const totalUsage = usageObj?.total ?? usageObj?.currentTurn;
-        toolCall.updateSubagentMetrics({
-          contextTokens: event.contextTokens,
-          usage: totalUsage,
-        });
-        return true;
-      }
-      case 'background.task.started':
-      case 'background.task.updated':
-      case 'background.task.terminated':
-      case 'compaction.blocked':
-      case 'compaction.cancelled':
-      case 'compaction.completed':
-      case 'compaction.started':
-      case 'error':
-      case 'session.meta.updated':
-      case 'skill.activated':
-      case 'subagent.completed':
-      case 'subagent.failed':
-      case 'subagent.spawned':
-      case 'tool.progress':
-      case 'tool.list.updated':
-      case 'mcp.server.status':
-      case 'turn.ended':
-      case 'turn.started':
-      case 'turn.step.completed':
-      case 'turn.step.interrupted':
-      case 'turn.step.retrying':
-      case 'turn.step.started':
-        return true;
-      default:
-        return true;
-    }
-  }
-
-  // Initializes turn-scoped buffers when the SDK starts a turn.
-  private handleTurnBegin(_event: TurnStartedEvent): void {
-    void _event;
-    this.resetLiveToolUiState();
-    this.state.currentStep = 0;
-    this.patchLivePane({
-      mode: 'waiting',
-      pendingApproval: null,
-      pendingQuestion: null,
-    });
-    this.setAppState({
-      isStreaming: true,
-      streamingPhase: 'waiting',
-      streamingStartTime: Date.now(),
-    });
-  }
-
-  // Finalizes turn-scoped state when the SDK completes a turn.
-  private handleTurnEnd(_event: TurnEndedEvent, sendQueued: (item: QueuedMessage) => void): void {
-    void _event;
-    this.flushStreamingUiUpdatesNow();
-    const todos = this.state.todoPanel.getTodos();
-    if (todos.length > 0 && todos.every((t) => t.status === 'done')) {
-      this.setTodoList([]);
-    }
-    this.resetLiveToolUiState();
-    this.finalizeTurn(sendQueued);
-  }
-
-  // Resets live render state for a new turn step.
-  private handleStepBegin(event: TurnStepStartedEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    this.state.currentStep = event.step;
-    this.resetLiveToolUiState();
-    this.finalizeLiveTextBuffers('waiting');
-    this.patchLivePane({
-      mode: 'waiting',
-      pendingApproval: null,
-      pendingQuestion: null,
-    });
-    this.setAppState({
-      streamingPhase: 'waiting',
-      streamingStartTime: Date.now(),
-    });
-  }
-
-  // Surfaces step-level outcomes the user needs to act on. The common
-  // case (finishReason === 'tool_use' or 'end_turn') is silent — those
-  // already render via tool.call.started/tool.result and assistant.delta.
-  // The interesting case is max_tokens: the model started a tool_use but
-  // ran out of budget before finalizing it, so the partial tool call is
-  // still pinned in 'Preparing' state with no signal that anything went
-  // wrong. Flip those into a visible 'Truncated' state and append a
-  // notice pointing at the config knob.
-  private handleStepCompleted(event: TurnStepCompletedEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    if (event.finishReason !== 'max_tokens') return;
-
-    // Scope the truncation marking to tool calls that belong to the
-    // step that just completed. Without this guard, stale entries from
-    // earlier retry attempts (or unrelated still-tracked calls) would
-    // get relabeled and counted, producing misleading "tool call was
-    // truncated" notices for the wrong step.
-    const eventTurnId = String(event.turnId);
-    let truncatedCount = 0;
-    for (const toolCall of this.state.activeToolCalls.values()) {
-      if (toolCall.result !== undefined) continue;
-      if (toolCall.streamingArguments === undefined) continue;
-      if (toolCall.turnId !== eventTurnId) continue;
-      if (toolCall.step !== event.step) continue;
-      toolCall.truncated = true;
-      const component = this.state.pendingToolComponents.get(toolCall.id);
-      if (component !== undefined) {
-        component.updateToolCall(toolCall);
-      }
-      truncatedCount += 1;
-    }
-    this.state.streamingToolCallArguments.clear();
-
-    const title =
-      truncatedCount > 0
-        ? 'Model hit max_tokens — tool call was truncated before it could run.'
-        : 'Model hit max_tokens — no tool call was emitted.';
-    // The `max_output_size` knob is only wired through to provider
-    // requests for the Anthropic provider (see toKosongProviderConfig).
-    // For OpenAI / Kimi / Google sessions the advice would be a
-    // dead end, so skip the second line on those providers.
-    const detail = this.isAnthropicSessionActive()
-      ? 'If this limit is wrong for your model, set `max_output_size` on the model alias in your kimi-code config.'
-      : undefined;
-    this.showNotice(title, detail);
-  }
-
-  private isAnthropicSessionActive(): boolean {
-    const providerKey = this.state.appState.availableModels[this.state.appState.model]?.provider;
-    if (providerKey === undefined) return false;
-    return this.state.appState.availableProviders[providerKey]?.type === 'anthropic';
-  }
-
-  // Renders user-facing status for an interrupted turn step.
-  private handleStepInterrupted(event: TurnStepInterruptedEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    this.resetLiveToolUiState();
-    this.finalizeLiveTextBuffers('idle');
-    const reason = event.reason;
-    if (reason === 'error') return;
-    if (reason === 'aborted' || reason === undefined || reason === '') {
-      this.showStatus('Interrupted by user', this.state.theme.colors.error);
-      return;
-    }
-    this.showError(
-      reason === 'max_steps'
-        ? 'reached per-turn step limit (max_steps)'
-        : `step interrupted (${reason})`,
-    );
-  }
-
-  // Appends a thinking delta to the live thinking block.
-  private handleThinkingDelta(event: ThinkingDeltaEvent): void {
-    this.state.thinkingDraft += event.delta;
-    this.streamingUI.markThinkingDirty();
-    this.patchLivePane({ mode: 'idle' });
-    if (this.state.appState.streamingPhase !== 'thinking') {
-      this.setAppState({ streamingPhase: 'thinking', streamingStartTime: Date.now() });
-    }
-    this.scheduleStreamingUiFlush();
-  }
-
-  // Appends an assistant text delta to the live assistant block.
-  private handleAssistantDelta(event: AssistantDeltaEvent): void {
-    if (this.state.thinkingDraft.length > 0) {
-      this.flushThinkingToTranscript('idle');
-    }
-
-    if (!this.state.assistantStreamActive) {
-      this.state.assistantStreamActive = true;
-      this.onStreamingTextStart();
-    }
-
-    this.state.assistantDraft += event.delta;
-    this.streamingUI.markAssistantDirty();
-
-    this.patchLivePane({
-      mode: 'idle',
-      pendingApproval: null,
-      pendingQuestion: null,
-    });
-    if (this.state.appState.streamingPhase !== 'composing') {
-      this.setAppState({ streamingPhase: 'composing', streamingStartTime: Date.now() });
-    }
-    this.scheduleStreamingUiFlush();
-  }
-
-  private handleHookResult(event: HookResultEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    if (this.state.thinkingDraft.length > 0) {
-      this.flushThinkingToTranscript('idle');
-    }
-    this.finalizeAssistantStream();
-    this.appendTranscriptEntry({
-      id: nextTranscriptId(),
-      kind: 'assistant',
-      turnId: String(event.turnId),
-      renderMode: 'markdown',
-      content: formatHookResultMarkdown(event),
-    });
-    this.patchLivePane({
-      mode: 'idle',
-      pendingApproval: null,
-      pendingQuestion: null,
-    });
-  }
-
-  // Starts or updates a rendered tool call from a tool-call start event.
-  private handleToolCall(event: ToolCallStartedEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    const toolCall: ToolCallBlockData = {
-      id: event.toolCallId,
-      name: event.name,
-      args: argsRecord(event.args),
-      description: event.description,
-      display: event.display,
-      step: this.state.currentStep,
-      turnId: this.state.currentTurnId,
-    };
-    const existing = this.state.activeToolCalls.get(event.toolCallId);
-    this.state.activeToolCalls.set(event.toolCallId, toolCall);
-    this.streamingUI.pendingToolCallFlushIds.delete(event.toolCallId);
-    this.state.streamingToolCallArguments.delete(event.toolCallId);
-    const existingComponent = this.state.pendingToolComponents.get(event.toolCallId);
-    if (existingComponent !== undefined) {
-      existingComponent.updateToolCall(toolCall);
-    } else if (existing === undefined) {
-      this.finalizeLiveTextBuffers('tool');
-      if (event.name !== 'Agent') {
-        this.onToolCallStart(toolCall);
-      }
-    }
-    this.patchLivePane({
-      mode: 'tool',
-      pendingApproval: null,
-      pendingQuestion: null,
-    });
-  }
-
-  // Accumulates streaming tool-call arguments and updates the rendered call.
-  private handleToolCallDelta(event: ToolCallDeltaEvent): void {
-    if (event.toolCallId.length === 0) return;
-    const id = event.toolCallId;
-    const existing = this.state.streamingToolCallArguments.get(id);
-    const argumentsText = appendStreamingArgsPreview(
-      existing?.argumentsText,
-      event.argumentsPart,
-    );
-    const name = event.name ?? existing?.name ?? this.state.activeToolCalls.get(id)?.name ?? 'Tool';
-    const startedAtMs = existing?.startedAtMs ?? Date.now();
-    this.state.streamingToolCallArguments.set(id, { name, argumentsText, startedAtMs });
-    this.streamingUI.pendingToolCallFlushIds.add(id);
-
-    this.patchLivePane({
-      mode: 'tool',
-      pendingApproval: null,
-      pendingQuestion: null,
-    });
-    if (this.state.appState.streamingPhase !== 'composing') {
-      this.setAppState({ streamingPhase: 'composing', streamingStartTime: Date.now() });
-    }
-    this.scheduleStreamingUiFlush();
-  }
-
-  // Streams a `{kind:'status'}` progress text into the live tool box so
-  // long-blocking tools (e.g. the MCP synthetic `authenticate` tool whose
-  // 15-minute browser wait would otherwise show only a spinner) can surface
-  // their authorization URL. Non-status update kinds stay out of the terminal
-  // transcript because only status text needs persistent display.
-  private handleToolProgress(event: ToolProgressEvent): void {
-    if (event.update.kind !== 'status') return;
-    const text = event.update.text;
-    if (text === undefined || text.length === 0) return;
-    const tc = this.state.pendingToolComponents.get(event.toolCallId);
-    if (tc === undefined) return;
-    tc.appendProgress(text);
-  }
-
-  // Completes a tool call and applies any tool-specific UI side effects.
-  private handleToolResult(event: ToolResultEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    const matchedCall = this.state.activeToolCalls.get(event.toolCallId);
-    const resultData: ToolResultBlockData = {
-      tool_call_id: event.toolCallId,
-      output: serializeToolResultOutput(event.output),
-      is_error: event.isError,
-      synthetic: event.synthetic,
-    };
-    if (matchedCall !== undefined) {
-      this.onToolCallEnd(event.toolCallId, resultData);
-      if (matchedCall.name === 'TodoList' && !event.isError) {
-        const rawTodos = (matchedCall.args as { todos?: unknown }).todos;
-        if (Array.isArray(rawTodos)) {
-          const sanitized = rawTodos
-            .filter((todo): todo is { title: string; status: 'pending' | 'in_progress' | 'done' } =>
-              isTodoItemShape(todo),
-            )
-            .map((t) => ({ title: t.title, status: t.status }));
-          this.setTodoList(sanitized);
-        }
-      }
-    }
-    this.state.activeToolCalls.delete(event.toolCallId);
-    this.state.streamingToolCallArguments.delete(event.toolCallId);
-    this.patchLivePane({ mode: 'waiting' });
-  }
-
-  // Applies agent status updates to app state.
-  private handleStatusUpdate(event: AgentStatusUpdatedEvent): void {
-    const patch: Partial<AppState> = {};
-    if (event.contextUsage !== undefined) patch.contextUsage = event.contextUsage;
-    if (event.contextTokens !== undefined) patch.contextTokens = event.contextTokens;
-    if (event.maxContextTokens !== undefined) patch.maxContextTokens = event.maxContextTokens;
-    if (event.planMode !== undefined) patch.planMode = event.planMode;
-    if (event.permission !== undefined) {
-      patch.permissionMode = event.permission;
-      patch.yolo = event.permission === 'yolo';
-    }
-    if (event.model !== undefined) patch.model = event.model;
-    if (Object.keys(patch).length > 0) this.setAppState(patch);
-  }
-
-  // Applies session metadata changes to the UI and process title.
-  private handleSessionMetaChanged(event: SessionMetaUpdatedEvent): void {
-    const title = event.title ?? stringValue(event.patch?.['title']);
-    if (title !== undefined) {
-      this.setAppState({ sessionTitle: title });
-      setProcessTitle(title, this.state.appState.sessionId);
-    }
-  }
-
-  // Finalizes live buffers and renders a session error.
-  private handleSessionError(event: ErrorEvent): void {
-    this.flushStreamingUiUpdatesNow();
-    this.resetLiveToolUiState();
-    this.finalizeLiveTextBuffers('idle');
-    if (event.code === OAUTH_LOGIN_REQUIRED_CODE) {
-      this.showError(OAUTH_LOGIN_REQUIRED_STARTUP_NOTICE);
-      return;
-    }
-    this.showError(`[${event.code}] ${event.message}`);
-    const sessionId = this.state.appState.sessionId;
-    if (sessionId.length > 0) {
-      this.showStatus(errorReportHintLine(sessionId));
-    }
-  }
-
-  private handleSessionWarning(event: WarningEvent): void {
-    this.showStatus(`Warning: ${event.message}`, this.state.theme.colors.warning);
-  }
-
-  private renderMcpServerStatus(server: McpServerStatusSnapshot): void {
-    const key = mcpServerStatusKey(server);
-    if (this.state.renderedMcpServerStatusKeys.get(server.name) === key) return;
-    this.state.renderedMcpServerStatusKeys.set(server.name, key);
-
-    const colors = this.state.theme.colors;
-    switch (server.status) {
-      case 'connected': {
-        const toolStr = `${server.toolCount} tool${server.toolCount === 1 ? '' : 's'}`;
-        const message = `MCP server "${server.name}" connected · ${toolStr} (${server.transport})`;
-        this.finalizeMcpServerStatusRow(server.name, message, colors.success);
-        return;
-      }
-      case 'failed': {
-        const message = `MCP server "${server.name}" failed${server.error !== undefined ? `: ${server.error}` : ''}`;
-        this.finalizeMcpServerStatusRow(server.name, message, colors.error);
-        return;
-      }
-      case 'needs-auth': {
-        const message = `MCP server "${server.name}" needs OAuth — run /mcp-config login ${server.name}`;
-        this.finalizeMcpServerStatusRow(server.name, message, colors.warning);
-        return;
-      }
-      case 'disabled':
-        this.finalizeMcpServerStatusRow(
-          server.name,
-          `MCP server "${server.name}" disabled`,
-          colors.textMuted,
-        );
-        return;
-      case 'pending':
-        this.showMcpServerStatusSpinner(server.name);
-        return;
-    }
-  }
-
-  private showMcpServerStatusSpinner(name: string): void {
-    const label = `MCP server "${name}" connecting…`;
-    const existing = this.state.mcpServerStatusSpinners.get(name);
-    if (existing !== undefined) {
-      existing.setLabel(label);
-      return;
-    }
-    const tint = (s: string): string => chalk.hex(this.state.theme.colors.textMuted)(s);
-    const spinner = new MoonLoader(this.state.ui, 'braille', tint, label);
-    this.state.transcriptContainer.addChild(spinner);
-    this.state.mcpServerStatusSpinners.set(name, spinner);
-    this.state.ui.requestRender();
-  }
-
-  private finalizeMcpServerStatusRow(name: string, message: string, color: string): void {
-    const spinner = this.state.mcpServerStatusSpinners.get(name);
-    if (spinner === undefined) {
-      this.showStatus(message, color);
-      return;
-    }
-    spinner.stop();
-    const status = new StatusMessageComponent(message, this.state.theme.colors, color);
-    const children = this.state.transcriptContainer.children;
-    const idx = children.indexOf(spinner);
-    if (idx >= 0) {
-      children[idx] = status;
-      this.state.transcriptContainer.invalidate();
-    } else {
-      this.state.transcriptContainer.addChild(status);
-    }
-    this.state.mcpServerStatusSpinners.delete(name);
-    this.state.ui.requestRender();
-  }
-
-  private stopAllMcpServerStatusSpinners(): void {
-    for (const spinner of this.state.mcpServerStatusSpinners.values()) {
-      spinner.stop();
-    }
-    this.state.mcpServerStatusSpinners.clear();
-  }
-
-  // Adds a skill activation entry to the transcript once.
-  private handleSkillActivated(event: SkillActivatedEvent): void {
-    if (this.state.renderedSkillActivationIds.has(event.activationId)) return;
-    this.state.renderedSkillActivationIds.add(event.activationId);
-    this.appendTranscriptEntry({
-      id: nextTranscriptId(),
-      kind: 'skill_activation',
-      turnId: undefined,
-      renderMode: 'plain',
-      content: `Activated skill: ${event.skillName}`,
-      skillActivationId: event.activationId,
-      skillName: event.skillName,
-      skillArgs: event.skillArgs,
-    });
-  }
-
-  // Starts the compaction UI block and marks the app as compacting.
-  private handleCompactionBegin(event: CompactionStartedEvent): void {
-    this.finalizeLiveTextBuffers('waiting');
-    this.setAppState({
-      isCompacting: true,
-      streamingPhase: 'waiting',
-      streamingStartTime: Date.now(),
-    });
-    this.beginCompaction(event.instruction);
-  }
-
-  // Finishes compaction and resumes queued work when possible.
-  private handleCompactionEnd(
-    event: CompactionCompletedEvent,
-    sendQueued: (item: QueuedMessage) => void,
-  ): void {
-    this.endCompaction(event.result.tokensBefore, event.result.tokensAfter);
-    this.finishCompaction(sendQueued);
-  }
-
-  private handleCompactionCancel(
-    _event: CompactionCancelledEvent,
-    sendQueued: (item: QueuedMessage) => void,
-  ): void {
-    this.cancelCompactionBlock();
-    this.finishCompaction(sendQueued);
-  }
-
-  private finishCompaction(sendQueued: (item: QueuedMessage) => void): void {
-    if (!this.state.appState.isStreaming) {
-      this.setAppState({
-        isCompacting: false,
-        streamingPhase: 'idle',
-      });
-      this.resetLivePane();
-      if (this.state.queuedMessages.length > 0) {
-        const [next, ...rest] = this.state.queuedMessages;
-        this.state.queuedMessages = rest;
-        if (next !== undefined) {
-          setTimeout(() => {
-            sendQueued(next);
-          }, 0);
-        }
-      }
-    } else {
-      this.setAppState({ isCompacting: false });
-    }
-  }
-
-  // Registers a spawned subagent and renders foreground or background status.
-  private handleSubagentSpawned(event: SubagentSpawnedEvent): void {
-    this.state.subagentParentToolCallIds.set(event.subagentId, event.parentToolCallId);
-    this.state.subagentNames.set(event.subagentId, event.subagentName);
-
-    if (event.runInBackground) {
-      const meta = this.buildBackgroundAgentMetadata(event);
-      this.state.backgroundAgentMetadata.set(event.subagentId, meta);
-      this.state.backgroundAgents.add(event.subagentId);
-      this.appendBackgroundAgentEntry('started', meta);
-      this.syncBackgroundAgentBadge();
-      return;
-    }
-
-    let tc = this.state.pendingToolComponents.get(event.parentToolCallId);
-    if (tc === undefined) {
-      const toolCall = this.state.activeToolCalls.get(event.parentToolCallId);
-      if (toolCall !== undefined) {
-        this.onToolCallStart(toolCall);
-        tc = this.state.pendingToolComponents.get(event.parentToolCallId);
-      }
-    }
-    tc ??= this.createStandaloneSubagentToolCall(event);
-    if (tc === undefined) return;
-    tc.onSubagentSpawned({
-      agentId: event.subagentId,
-      agentName: event.subagentName,
-      runInBackground: event.runInBackground,
-    });
-  }
-
-  // Completes a subagent in its parent tool call or background transcript entry.
-  private handleSubagentCompleted(event: SubagentCompletedEvent): void {
-    const backgroundMeta = this.state.backgroundAgentMetadata.get(event.subagentId);
-    if (this.state.backgroundAgents.delete(event.subagentId)) {
-      this.syncBackgroundAgentBadge();
-    }
-    if (backgroundMeta !== undefined) {
-      this.state.backgroundAgentMetadata.delete(event.subagentId);
-      // Dedupe: if the BPM `background.task.terminated` for the
-      // matching agent task already pushed a terminal card, skip.
-      // Otherwise mark the subagent id so a later BPM event skips.
-      const taskId = this.findAgentTaskId(event.subagentId);
-      if (taskId !== undefined && this.state.backgroundTaskTranscriptedTerminal.has(taskId)) {
-        return;
-      }
-      if (taskId !== undefined) {
-        this.state.backgroundTaskTranscriptedTerminal.add(taskId);
-      }
-      const extras =
-        event.resultSummary === undefined ? undefined : { resultSummary: event.resultSummary };
-      this.appendBackgroundAgentEntry('completed', backgroundMeta, extras);
-      return;
-    }
-    const tc = this.state.pendingToolComponents.get(event.parentToolCallId);
-    if (tc === undefined) return;
-    tc.onSubagentCompleted({
-      contextTokens: event.contextTokens,
-      usage: event.usage,
-      resultSummary: event.resultSummary,
-    });
-    if (!this.state.activeToolCalls.has(event.parentToolCallId)) {
-      this.state.pendingToolComponents.delete(event.parentToolCallId);
-    }
-  }
-
-  // Marks a subagent failure in its parent tool call or background transcript entry.
-  private handleSubagentFailed(event: SubagentFailedEvent): void {
-    const backgroundMeta = this.state.backgroundAgentMetadata.get(event.subagentId);
-    if (this.state.backgroundAgents.delete(event.subagentId)) {
-      this.syncBackgroundAgentBadge();
-    }
-    if (backgroundMeta !== undefined) {
-      this.state.backgroundAgentMetadata.delete(event.subagentId);
-      const taskId = this.findAgentTaskId(event.subagentId);
-      if (taskId !== undefined && this.state.backgroundTaskTranscriptedTerminal.has(taskId)) {
-        return;
-      }
-      if (taskId !== undefined) {
-        this.state.backgroundTaskTranscriptedTerminal.add(taskId);
-      }
-      this.appendBackgroundAgentEntry('failed', backgroundMeta, { error: event.error });
-      return;
-    }
-    const tc = this.state.pendingToolComponents.get(event.parentToolCallId);
-    if (tc === undefined) return;
-    tc.onSubagentFailed({ error: event.error });
-    if (!this.state.activeToolCalls.has(event.parentToolCallId)) {
-      this.state.pendingToolComponents.delete(event.parentToolCallId);
-    }
-  }
-
-  // Mounts subagents launched by session-level commands that do not originate
-  // from a model-issued Agent tool call.
-  private createStandaloneSubagentToolCall(event: SubagentSpawnedEvent): ToolCallComponent | undefined {
-    const description = event.description ?? `Run ${event.subagentName} agent`;
-    const toolCall: ToolCallBlockData = {
-      id: event.parentToolCallId,
-      name: 'Agent',
-      args: {
-        description,
-        subagent_type: event.subagentName,
-      },
-      description,
-      step: this.state.currentStep,
-      turnId: this.state.currentTurnId,
-    };
-    this.onToolCallStart(toolCall);
-    return this.state.pendingToolComponents.get(event.parentToolCallId);
-  }
-
-  /**
-   * Locate the BPM `agent-*` task id whose `description` matches the
-   * spawned subagent's recorded description. Used only for dedupe
-   * between the BPM and subagent flows — best-effort: if there is no
-   * unique match (e.g. multiple agent tasks with the same description)
-   * the caller treats the dedupe as a miss, which is safe.
-   */
-  private findAgentTaskId(subagentId: string): string | undefined {
-    const meta = this.state.backgroundAgentMetadata.get(subagentId);
-    const description = meta?.description ?? meta?.agentName;
-    if (description === undefined) return undefined;
-    let match: string | undefined;
-    for (const info of this.state.backgroundTasks.values()) {
-      if (!info.taskId.startsWith('agent-')) continue;
-      if (info.description !== description) continue;
-      if (match !== undefined) return undefined; // ambiguous
-      match = info.taskId;
-    }
-    return match;
-  }
-
-  // Builds transcript metadata for a background subagent.
-  private buildBackgroundAgentMetadata(event: SubagentSpawnedEvent): BackgroundAgentMetadata {
-    const parent = this.state.activeToolCalls.get(event.parentToolCallId);
-    const description = parent?.args['description'] ?? event.description;
-    return {
-      agentId: event.subagentId,
-      parentToolCallId: event.parentToolCallId,
-      agentName: event.subagentName,
-      description: typeof description === 'string' ? description : undefined,
-    };
-  }
-
-  // Appends a background-agent status row to the transcript.
-  private appendBackgroundAgentEntry(
-    phase: 'started' | 'completed' | 'failed',
-    meta: BackgroundAgentMetadata,
-    extras: { resultSummary?: string; error?: string } | undefined = undefined,
-  ): void {
-    const status = formatBackgroundAgentTranscript(phase, meta, extras);
-    const entry: TranscriptEntry = {
-      id: nextTranscriptId(),
-      kind: 'status',
-      turnId: this.state.currentTurnId,
-      renderMode: 'plain',
-      content: status.headline,
-      detail: status.detail,
-      backgroundAgentStatus: status,
-    };
-    this.appendTranscriptEntry(entry);
-  }
-
-  // Updates the footer badge for active background agents.
-  private syncBackgroundAgentBadge(): void {
-    this.syncBackgroundTaskBadge();
-  }
-
-  // =========================================================================
-  // Background task lifecycle (BPM-derived, covers both bash + agent tasks)
-  // =========================================================================
-
-  private handleBackgroundTaskEvent(
-    event: BackgroundTaskStartedEvent | BackgroundTaskUpdatedEvent | BackgroundTaskTerminatedEvent,
-  ): void {
-    const { info } = event;
-    const previous = this.state.backgroundTasks.get(info.taskId);
-    this.state.backgroundTasks.set(info.taskId, info);
-
-    // If the user is currently viewing this task's output, nudge a
-    // refresh immediately so they see new content without waiting for
-    // the 1s poll. Same dedupe-by-output-equality applies inside.
-    const viewer = this.state.tasksBrowser?.viewer;
-    if (viewer !== undefined && viewer.taskId === info.taskId) {
-      void this.refreshTaskOutputViewer({ silent: true });
-    }
-
-    const isTerminal =
-      info.status === 'completed' ||
-      info.status === 'failed' ||
-      info.status === 'killed' ||
-      info.status === 'lost';
-
-    if (event.type === 'background.task.started') {
-      // For agent-* tasks, the legacy subagent.spawned flow already
-      // pushed a 'started' transcript card; skip to avoid duplicates.
-      if (info.taskId.startsWith('agent-')) {
-        this.syncBackgroundTaskBadge();
-        this.repaintTasksBrowser();
-        return;
-      }
-      this.appendBackgroundTaskEntry(info);
-      this.syncBackgroundTaskBadge();
-      this.repaintTasksBrowser();
-      return;
-    }
-
-    if (event.type === 'background.task.terminated' && isTerminal) {
-      if (!this.state.backgroundTaskTranscriptedTerminal.has(info.taskId)) {
-        // For agent-* tasks, the older subagent.completed/failed flow
-        // may also produce a terminal card; whoever wins records the
-        // dedupe marker first. See handleSubagentCompleted/Failed.
-        if (info.taskId.startsWith('bash-')) {
-          this.appendBackgroundTaskEntry(info);
-        }
-        this.state.backgroundTaskTranscriptedTerminal.add(info.taskId);
-      }
-      this.syncBackgroundTaskBadge();
-      this.repaintTasksBrowser();
-      return;
-    }
-
-    // updated: status flipped between running and awaiting_approval.
-    // No transcript card — just sync the badge if the active count
-    // changed (awaiting_approval still counts as active).
-    if (previous?.status !== info.status) {
-      this.syncBackgroundTaskBadge();
-    }
-    this.repaintTasksBrowser();
-  }
-
-  private appendBackgroundTaskEntry(info: BackgroundTaskInfo): void {
-    const status = formatBackgroundTaskTranscript(info);
-    const entry: TranscriptEntry = {
-      id: nextTranscriptId(),
-      kind: 'status',
-      turnId: this.state.currentTurnId,
-      renderMode: 'plain',
-      content: status.headline,
-      detail: status.detail,
-      backgroundAgentStatus: status,
-    };
-    this.appendTranscriptEntry(entry);
-  }
-
-  // Footer counts are BPM-derived: every task that is not terminal,
-  // split by id prefix so bash and agent badges render independently.
-  // awaiting_approval still counts as active; lost/killed/completed/
-  // failed do not.
-  private syncBackgroundTaskBadge(): void {
-    let bashTasks = 0;
-    let agentTasks = 0;
-    for (const info of this.state.backgroundTasks.values()) {
-      if (
-        info.status === 'completed' ||
-        info.status === 'failed' ||
-        info.status === 'killed' ||
-        info.status === 'lost'
-      ) {
-        continue;
-      }
-      if (info.taskId.startsWith('agent-')) {
-        agentTasks += 1;
-      } else {
-        bashTasks += 1;
-      }
-    }
-    this.state.footer.setBackgroundCounts({ bashTasks, agentTasks });
-    this.state.ui.requestRender();
-  }
+  private stopAllMcpServerStatusSpinners(): void { this.sessionEventHandler.stopAllMcpServerStatusSpinners(); }
 
   // =========================================================================
   // Live Render Hooks — delegated to StreamingUIController
@@ -3560,7 +2608,7 @@ export class KimiTUI {
   }
 
   // Stores a transcript entry and mounts its component if renderable.
-  private appendTranscriptEntry(entry: TranscriptEntry): void {
+  appendTranscriptEntry(entry: TranscriptEntry): void {
     this.state.transcriptEntries.push(entry);
     const component = this.createTranscriptComponent(entry);
     if (component) {
@@ -3648,7 +2696,7 @@ export class KimiTUI {
   }
 
   // Appends a status message to the transcript.
-  private showStatus(message: string, color?: string): void {
+  showStatus(message: string, color?: string): void {
     this.state.transcriptContainer.addChild(
       new StatusMessageComponent(message, this.state.theme.colors, color),
     );
@@ -3656,7 +2704,7 @@ export class KimiTUI {
   }
 
   // Appends a notice message to the transcript.
-  private showNotice(title: string, detail?: string): void {
+  showNotice(title: string, detail?: string): void {
     this.state.transcriptContainer.addChild(
       new NoticeMessageComponent(title, detail, this.state.theme.colors),
     );
@@ -4058,8 +3106,8 @@ export class KimiTUI {
 
   private showTasksBrowser(): Promise<void> { return this.tasksBrowserController.show(); }
   private closeTasksBrowser(): void { this.tasksBrowserController.close(); }
-  private repaintTasksBrowser(): void { this.tasksBrowserController.repaint(); }
-  private refreshTaskOutputViewer(opts?: { silent?: boolean }): Promise<void> {
+  repaintTasksBrowser(): void { this.tasksBrowserController.repaint(); }
+  refreshTaskOutputViewer(opts?: { silent?: boolean }): Promise<void> {
     return this.tasksBrowserController.refreshOutputViewer(opts);
   }
